chore: Add a getter for the tool's input schema (#30)
* chore: Add support for optional parameters

* change error message

* minor fix

* chore: Add cloud build reporter for tests (#19)

* chore: Update sync-labels.yaml (#22)

Minor fix in the name of the labels.yaml file in the workflow

* chore: Update README.md (#23)

Minor change to the command to download the sdk

* chore: Seperate E2E tests from unit tests (#28)

* chore: Seperate E2E tests from unit tests

This PR seperates the E2E tests from unit tests by using build tags.

This change will make sure the setup for E2E tests is not run for the unit tests.

* Update e2e_test.go

* Update integration.cloudbuild.yaml

* Update client_test.go

* Update options_test.go

* Update protocol_test.go

* Update tool_test.go

* Update utils_test.go

* Update integration.cloudbuild.yaml

* Update integration.cloudbuild.yaml

* chore: Add better assert condition for a flaky test (#27)

* chore: Add better assert condition for a flaky test

This particular test was flaky as the loadToolset runs an iteration through the map of Tools and the map iteration order is intentionally randomized.

This assert condition will correctly catch the error even if the map iteration order is random.

* minor fix

* Update client_test.go

* chore: Add DefaultToolOptions usage in the readme (#29)

* chore: Add DefaultToolOptions usage in the readme

This change adds the usage of DefaultToolOptions in the documentation.

* Update README.md

* Update README.md

* chore: Add GetGoogleIDToken method (#24)

* chore: Add GetGoogleIDToken method

* Use GetGoogleIDToken in the E2E test

* Remove GetGoogleIDToken in the E2E test

* Add documentation for GetGoogleIDToken

* chore: Add a getter for the tool's input schema

* add docstrings

* Update core/tool.go

Co-authored-by: Kurtis Van Gent <31518063+kurtisvg@users.noreply.github.com>

---------

Co-authored-by: Kurtis Van Gent <31518063+kurtisvg@users.noreply.github.com>